[C] Move useCollapseContext to component dir
