+++ client/src/hooks/annotationHelpers.js
